react-flask-app/server/requirements.txt
alembic==1.13.1
asn1crypto==1.5.1
bcrypt==4.1.3
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
cbor2==5.6.4
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-mailman==1.1.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-SeaSurf==1.1.1
Flask-Security-Too==5.4.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
passlib==1.7.4
phonenumberslite==8.13.39
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
six==1.16.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
webauthn==2.1.0
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
